+++ src/i18n/LocaleSwitcher.vue
@@ -1,6 +1,8 @@
 <script setup lang="ts">
 import { PhGlobe } from "@phosphor-icons/vue";
-import { languageNames } from "./i18n";
+import { useAppConfig } from "@/app/useAppConfig";
+
+const { uiLanguages } = useAppConfig();
 </script>
 
 <template>
@@ -11,12 +13,8 @@ import { languageNames } from "./i18n";
       v-model="$i18n.locale"
       class="bg-transparent border-0 appearance-none py-0 ps-4 pe-3 -ms-5 w-7 cursor-pointer"
     >
-      <option
-        v-for="locale in $i18n.availableLocales"
-        :key="locale"
-        :value="locale"
-      >
-        {{ languageNames[locale] }}
+      <option v-for="{ id, key, name } in uiLanguages" :key :value="id">
+        {{ name }}
       </option>
     </select>
   </div>